Kill `Fn` impls for `DynamicStyle`
It appears this is a "Crust of Rust" type of scenario. You can't implement (generically) multiple `Fn` signatures to have the same trait. Which is really odd to me considering the functions will be different types, it just seems that the compiler is not smart enough to figure that out? Anyways, here's a playground for that issue: https://play.rust-lang.org/?version=stable&mode=debug&edition=2021&gist=9e1d9bfc8ace783182802f761d044c48.